Bug Fix
A failed edge conditional with a passed ordering included the node.
Fix was to make sure the nodeHash is either missing the node in the first
place (previous code) OR the value is None, meaning it failed the
conditional test.  We've had this bug for years!

Anoop found the bug but I still have CVS access