Before setting the IP address of an interface check that it is valid
This fix is long overdue, basically if the user creates an interface
with an invalid IP address then it is unable to remove the host (without,
removing the interface before).

This solution is not optimal (although better than before) because if you
run rocks add interface with an invalid value it stil creates an empty
interface (but now it does not insert the invalid IP address).