devel: user docs now handles properly the ROLL_REQUIRES_FULL_OS variable
all the user docs where generated with the warning:
---------
This roll has been released independent of the corresponding Rocks release.
It therefore requires the complete OS roll and will not function correctly
if using only the Jumbo or incomplete set of OS CDROMs.
---------

Now only if you define the variable ROLL_REQUIRES_FULL_OS the warning
will be generated