pylib: add include attribute to file tag in node.xml
This commit add support for the include attribute to the
file tag. In this way it is possible to include file on
the frontend node without using the eval tag.
See the discussion on the ML:
http://marc.info/?l=npaci-rocks-discussion&m=135938905904484&w=2

Some example:

    <file name="/etc/newfile" include="/etc/oldfile"/>

Copy the content of /etc/oldfile inside /etc/newfile
cat /etc/oldfile > /etc/newfile
Where oldfile is from the frontend.

   <file name="/etc/newfile" include="/etc/oldfile" mode="append"/>

Append the content of /etc/oldfile inside /etc/newfile
Where oldfile is from the frontend and newfile is on the installing host