kickstart: removed code to insert new nodes in kickstart
This patch remove a chunk of code which was inserting nodes into
the database even though the node did not already exist in the DB.

Normally nodes are already present in the DB when the try to
kickstart, either because they were discovered with insert-ethers
or because they are simply reinstalling.

If this code does something usefull, this is not documented in
anywhere.